requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
passlib==1.7.4
mysqlclient==2.2.0
